requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
pydantic>=2.5
python-multipart>=0.0.6
tqdm>=4.66

# ASR engines (choose at least one)
faster-whisper>=0.9.0
openai-whisper>=20231116

# Optional for MSLite engine
transformers>=4.40
numpy>=1.26